pipelines/gemini-analysis/requirements.txt
functions-framework==3.*
google-generativeai>=0.8.0
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.16.0
